requirements.txt
# Project dependencies
torch==2.9.1
torchvision==0.24.1
numpy==2.4.0
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.16.3
pillow==12.1.0
matplotlib==3.10.8
PyYAML==6.0.3
streamlit>=1.40.0
